GATES

+++ app/Http/Controllers/EditorasController.php
@@ -1,6 +1,7 @@
 <?php
 
 namespace App\Http\Controllers;
+use Illuminate\Support\Facades\Gate;
 use Auth;
 use Illuminate\Http\Request;
 use App\Models\Editora;
@@ -15,6 +16,9 @@ class EditorasController extends Controller
             'editoras'=>$editoras
         ]);
     }
+
+
+
     public function show(Request $request){
         $idEditora = $request->ide;
         //$editora=Editora::findOrFail($idEditora);
@@ -24,25 +28,55 @@ class EditorasController extends Controller
             'editora'=>$editora
         ]);
     }
+
+
+
     public function create(){
-        return view('editoras.create');
+        if(Gate::allows('admin')){
+            return view('editoras.create');
+        }
+        else{
+            return redirect()->route('editoras.index')
+                ->with('msg','Não têm permissão para aceder a area pretendida');
+        }
     }
+
+
+
     public function store(Request $request){
         $novoEditora=$request->validate([
             'nome'=>['required','min:3','max:100'],
             'morada'=>['nullable','min:3','max:255'],
             'observacoes'=>['nullable','min:3','max:255']
         ]);
-        $editora=Editora::create($novoEditora);
-        return redirect()->route('editoras.show',[
-            'ide'=>$editora->id_editora
-        ]);
+        if(Gate::allows('admin')){
+            $editora=Editora::create($novoEditora);
+            return redirect()->route('editoras.show',[
+                'ide'=>$editora->id_editora
+            ]);
+        }
+        else{
+            return redirect()->route('editoras.index')
+                ->with('msg','Não têm permissão para aceder a area pretendida');
+        }
     }
+
+
+
     public function edit(Request $request){
         $ide = $request->ide;
         $editora=Editora::where('id_editora',$ide)->with('livros')->first();
-        return view('editoras.edit',['editora'=>$editora]);
+        if(Gate::allows('admin')){
+            return view('editoras.edit',['editora'=>$editora]);
+        }
+        else{
+            return redirect()->route('editoras.index')
+                ->with('msg','Não têm permissão para aceder a area pretendida');
+        }
     }
+
+
+
     public function update(Request $request){
         $ide = $request->ide;
         $editora=Editora::where('id_editora',$ide)->with('livros')->first();
@@ -51,30 +85,54 @@ class EditorasController extends Controller
             'morada'=>['nullable','min:3','max:255'],
             'observacoes'=>['nullable','min:3','max:255']
         ]);
-        $editarEditora=$editora->update($editEditora);
-        return redirect()->route('editoras.show',[
-            'ide'=>$editora->id_editora
-        ]);
+        if(Gate::allows('admin')){
+            $editarEditora=$editora->update($editEditora);
+            return redirect()->route('editoras.show',[
+                'ide'=>$editora->id_editora
+            ]);
+        }
+        else{
+            return redirect()->route('editoras.index')
+                ->with('msg','Não têm permissão para aceder a area pretendida');
+        }
     }
+
+
+
     public function deleted(Request $r){
         $ide = $r->ide;
         $editora=Editora::where('id_editora',$ide)->first();
-        if(is_null($editora)){
-            return redirect()->route('editoras.index')->with('msg','Não existe esta editora');
+        if(Gate::allows('admin')){
+            if(is_null($editora)){
+                return redirect()->route('editoras.index')->with('msg','Não existe esta editora');
+            }
+            else{
+                return view('editoras.delete',['editora'=>$editora]);
+            }
         }
         else{
-            return view('editoras.delete',['editora'=>$editora]);
+            return redirect()->route('editoras.index')
+                ->with('msg','Não têm permissão para aceder a area pretendida');
         }
     }
+
+
+
     public function destroy(Request $r){
         $ide = $r->ide;
         $editora=Editora::where('id_editora',$ide)->first();
-        if(is_null($editora)){
-            return redirect()->route('editoras.index')->with('msg','Não existe esta editora');
+        if(Gate::allows('admin')){
+            if(is_null($editora)){
+                return redirect()->route('editoras.index')->with('msg','Não existe esta editora');
+            }
+            else{
+                $eliminareditora=$editora->delete();
+                return redirect()->route('editoras.index');
+            }
         }
         else{
-            $eliminareditora=$editora->delete();
-            return redirect()->route('editoras.index');
+            return redirect()->route('editoras.index')
+                ->with('msg','Não têm permissão para aceder a area pretendida');
         }
     }
 }
